Automatic fix reference path in output .d.ts files
For https://github.com/kotas/gulp-tsc/issues/26

In output .d.ts files, tsc used a temporary directory as a base for
<reference> paths, which is wrong when the output directory does not
have the same level as the temporary directory.

To fix this, we introduce an output filter to fix reference paths
for output .d.ts files only. Same as fixing sourcemaps, we need
`outDir` is specified to know destination paths.